% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.clustcif.R
\name{autoplot.clustcif}
\alias{autoplot.clustcif}
\title{Visualization of \code{clustcif} objects with ggplot2 graphics}
\usage{
\method{autoplot}{clustcif}(
  object = object,
  groups_by_colour = TRUE,
  centers = FALSE,
  conf.int = FALSE,
  xlab = "Time",
  ylab = "Cumulative Incidence Function",
  interactive = FALSE,
  ...
)
}
\arguments{
\item{object}{Object of \code{clustcif} class.}

\item{groups_by_colour}{A specification for the plotting groups by color.}

\item{centers}{Draw the centroids (mean of the curves pertaining to the
same group) into the plot. By default it is \code{FALSE}.}

\item{conf.int}{Only for survival curves. Logical flag indicating whether to
plot confidence intervals. #conf.int = 0.95}

\item{xlab}{A title for the \code{x} axis.}

\item{ylab}{A title for the \code{y} axis.}

\item{interactive}{Logical flag indicating if an interactive plot with plotly is produced.}

\item{\ldots}{Other options. #ver si meter conf.type=c("log", "plain", "none"),}
}
\value{
A ggplot object, so you can use common features from
ggplot2 package to manipulate the plot.
}
\description{
Useful for drawing the estimated functions grouped by
color and the centroids (mean curve of the functions pertaining to the
same group).
}
\details{
See help page of the function \code{\link{autoplot.clustcif}}.
}
\examples{
\dontrun{
library(clustcif)
library(mstate)
data("ebmt2")

# Testing the number of groups
outtest <- clustcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                        kvector = NULL,  kbin = 20, nboot = 100,
                        algorithm = "kmeans", cluster = TRUE)
autoplot(outtest)
}
}
\author{
Nora M. Villanueva and Marta Sestelo.
}
