% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcif.R
\name{clustcif}
\alias{clustcif}
\title{Clustering Cumulative Incidence Functions in competing risks framework}
\usage{
clustcif(
  time,
  status = NULL,
  x,
  max_time = NULL,
  labels = NULL,
  kvector = NULL,
  kbin = 50,
  weights = NULL,
  nboot = 100,
  algorithm = "kmeans",
  alpha = 0.05,
  cluster = TRUE,
  ncores = NULL,
  seed = NULL
)
}
\arguments{
\item{time}{A numeric vector containing the follow up time.}

\item{status}{Event type indicator. Each value corresponds to the cause of the event.
A value of \code{0} indicates that the observation is censored (i.e., still
alive or event–free at the end of follow-up), and positive integers
(\code{1, 2, ...}) represent different event types (causes of failure).}

\item{x}{Categorical variable indicating the population to which
the observations belongs.}

\item{max_time}{Maximum time horizon used to compute the
cumulative incidence functions. If \code{NULL}, the maximum observed time is used.}

\item{labels}{A character vector providing the labels for the event types encoded in
\code{status}, including the label for censored observations. The vector may
have the same length as \code{status}, in which case it will be interpreted
as an observation-level vector and internally converted to unique event labels.
The number of distinct labels must match the number of event
categories present in \code{status}.}

\item{kvector}{A vector specifying the number of groups of curves to be
checking.}

\item{kbin}{Size of the grid over which the cumulative incidence functions
are to be estimated.}

\item{weights}{Type of weighting to use when computing the test statistic.
Options are:
\itemize{
\item{\code{NULL} (default): equal weights at all time points.}
\item{\code{"KM"}: Kaplan–Meier–based weights, smoothed and rescaled,
giving more weight to early times.}
\item{\code{"var"}: variance-based weights, inversely proportional to
CIF standard errors, giving more weight to precise estimates.}
}}

\item{nboot}{Number of bootstrap repeats.}

\item{algorithm}{A character string specifying which clustering algorithm is used,
i.e., k-means(\code{"kmeans"}) or k-medians (\code{"kmedians"}).}

\item{alpha}{Significance level of the testing procedure. Defaults to 0.05.}

\item{cluster}{A logical value. If  \code{TRUE} (default), the
testing procedure is  parallelized. Note that there are cases
(e.g., a low number of bootstrap repetitions) that R will gain in
performance through serial computation. R takes time to distribute tasks
across the processors also it will need time for binding them all together
later on. Therefore, if the time for distributing and gathering pieces
together is greater than the time need for single-thread computing, it does
not worth parallelize.}

\item{ncores}{An integer value specifying the number of cores to be used
in the parallelized procedure. If \code{NULL} (default), the number of cores
to be used is equal to the number of cores of the machine - 1.}

\item{seed}{Seed to be used in the procedure.}
}
\value{
A list containing the following items:
\item{table}{A data frame containing the null hypothesis tested, the values
of the test statistic and the obtained pvalues.}
\item{levels}{Original levels of the variable \code{x}.}
\item{cluster}{A vector of integers (from 1:k) indicating the cluster to
which each curve is allocated.}
\item{centers}{An object containing the centroids
(mean of the curves pertaining to the same group).}
\item{curves}{An object containing the fitted curves for each population.}
}
\description{
Function for grouping cumulative incidence curves (CIF) based on
the k-means or k-medians algorithm. It returns the optimal number of
clusters and the assignment to the cluster which they belong to.
}
\examples{
\dontrun{
library(clustcif)
library(mstate)
data("ebmt2")

# Testing the number of groups
outtest <- clustcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                        kvector = NULL,  kbin = 20, nboot = 100,
                        algorithm = "kmeans", cluster = TRUE)
summary(outtest)
}
}
\author{
Nora M. Villanueva and Marta Sestelo.
}
