% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcif.R
\name{kcif}
\alias{kcif}
\title{k-groups of multiple Cumulative Incidents Functions (CIF)}
\usage{
kcif(
  time,
  status = NULL,
  x = NULL,
  max_time = NULL,
  labels = NULL,
  k,
  kbin = 50,
  weights = NULL,
  algorithm = "kmeans",
  seed = NULL
)
}
\arguments{
\item{time}{Survival time.}

\item{status}{Censoring indicator of the survival
time of the process; 0 if the total time is censored and 1 otherwise.}

\item{x}{Categorical variable indicating the population to which
the observations belongs.}

\item{max_time}{TODO}

\item{labels}{TODO}

\item{k}{An integer specifying the number of groups of curves to be
performed.}

\item{kbin}{Size of the grid over which the cumulative incidence functions
are to be estimated.}

\item{weights}{TODO}

\item{algorithm}{A character string specifying which clustering algorithm is used,
i.e., k-means(\code{"kmeans"}) or k-medians (\code{"kmedians"}).}

\item{seed}{Seed to be used in the procedure.}
}
\value{
A list containing the following items:
\item{measure}{Value of the test statistics.}
\item{levels}{Original levels of the variable \code{x}.}
\item{cluster}{A vector of integers (from 1:k) indicating the cluster to
which each curve is allocated.}
\item{centers}{An object of class \code{survfit} containing the centroids
(mean of the cif pertaining to the same group).}
\item{curves}{An object of class \code{survfit} containing the cumulative incidence
functions for each population.}
}
\description{
Function for grouping CIF, given a number k,
based on the k-means or k-medians clustering algorithms.
}
\examples{
library(mstate) # data from this package
library(clustcif)
data("ebmt2")
out <- table(ebmt2$cod) # causes of death
table9 <- c(out[-1], out[1]) # reorder causes of death
names(table9)[7] <- "Censored" # changed name of the last cause of death
(table9 <- data.frame(t(table9)))
#Estimating the CIFs given a number of groups k = 2
out <- kcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                 k = 2, kbin = 50, algorithm = "kmeans")

out

}
\author{
Nora M. Villanueva and Marta Sestelo.
}
