% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.clustcif.R
\name{summary.clustcif}
\alias{summary.clustcif}
\title{Summarizing fits of \code{clustcif} class produced by \code{clustcif}}
\usage{
\method{summary}{clustcif}(object, ...)
}
\arguments{
\item{object}{a clustccif object as producted by \code{clustcif}}

\item{\ldots}{additional arguments.}
}
\value{
\code{summary.clustcif} computes and returns a list of summary
information for a \code{clustcurves} object.
\item{levels}{Levels of the factor.}
\item{cluster}{A vector containing the assignment of each factor's level to its group.}
\item{table}{A data.frame containing the results from the hypothesis test.}
}
\description{
Takes a  clustcif object
and produces various useful summaries from it.
}
\details{
\code{print.clustcif} tries to be smart about \code{summary.clustcif}.
}
\examples{
library(clustcif)
library(mstate)
data("ebmt2")

# Testing the number of groups
outtest <- clustcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                        kvector = NULL,  kbin = 20, nboot = 1,
                        algorithm = "kmeans", cluster = FALSE)
summary(outtest)

}
\author{
Nora M. Villanueva and Marta Sestelo.
}
