% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.kcif.R
\name{summary.kcurves}
\alias{summary.kcurves}
\title{Summarizing fits of \code{kcif} class produced by \code{kcif}}
\usage{
\method{summary}{kcurves}(object, ...)
}
\arguments{
\item{object}{a kcif object as produced by \code{kcif}.}

\item{\ldots}{additional arguments.}
}
\value{
\code{summary.kcif} computes and returns a list of summary
information for a \code{kcif} object.
\item{levels}{Levels of the factor.}
\item{cluster}{A vector containing the assignment of each factor's level to its group.}
}
\description{
Takes a  kcif object
and produces various useful summaries from it.
}
\details{
\code{print.kcif} tries to be smart about \code{summary.kcif}.
}
\examples{
library(mstate) # data from this package
library(clustcif)
data("ebmt2")
out <- table(ebmt2$cod) # causes of death
table9 <- c(out[-1], out[1]) # reorder causes of death
names(table9)[7] <- "Censored" # changed name of the last cause of death
(table9 <- data.frame(t(table9)))
#Estimating the CIFs given a number of k = 2 groups
out <- kcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                 k = 2, kbin = 50, algorithm = "kmeans")

summary(out)



}
\author{
Nora M. Villanueva and Marta Sestelo.
}
