% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustcif.R
\name{clustcif}
\alias{clustcif}
\title{Clustering Cumulative Incidence Functions in competing risks framework}
\usage{
clustcif(
  time,
  status = NULL,
  x,
  max_time = NULL,
  labels = NULL,
  kvector = NULL,
  kbin = 50,
  weights = NULL,
  nboot = 100,
  algorithm = "kmeans",
  alpha = 0.05,
  cluster = TRUE,
  ncores = NULL,
  seed = NULL,
  multiple = FALSE,
  multiple.method = "holm"
)
}
\arguments{
\item{time}{A numeric vector containing the follow up time.}

\item{status}{A numeric vector containing a censoring indicator of the time of the process; \code{"0"} if the total time is censored (or the individual is alive),
\code{">0"} if different causes of death are observed (1,2,3,...).}

\item{x}{Categorical variable indicating the population to which
the observations belongs.}

\item{max_time}{TODO}

\item{labels}{A categorical variable indicating the different causes of death, i.e.,
the competing risks.}

\item{kvector}{A vector specifying the number of groups of curves to be
checking.}

\item{kbin}{Size of the grid over which the survival functions
are to be estimated.}

\item{weights}{NULL}

\item{nboot}{Number of bootstrap repeats.}

\item{algorithm}{A character string specifying which clustering algorithm is used,
i.e., k-means(\code{"kmeans"}) or k-medians (\code{"kmedians"}).}

\item{alpha}{Significance level of the testing procedure. Defaults to 0.05.}

\item{cluster}{A logical value. If  \code{TRUE} (default), the
testing procedure is  parallelized. Note that there are cases
(e.g., a low number of bootstrap repetitions) that R will gain in
performance through serial computation. R takes time to distribute tasks
across the processors also it will need time for binding them all together
later on. Therefore, if the time for distributing and gathering pieces
together is greater than the time need for single-thread computing, it does
not worth parallelize.}

\item{ncores}{An integer value specifying the number of cores to be used
in the parallelized procedure. If \code{NULL} (default), the number of cores
to be used is equal to the number of cores of the machine - 1.}

\item{seed}{Seed to be used in the procedure.}

\item{multiple}{A logical value. If  \code{TRUE} (not default), the resulted
pvalues are adjusted by using one of several methods for multiple comparisons.}

\item{multiple.method}{Correction method. See Details.}
}
\value{
A list containing the following items:
\item{table}{A data frame containing the null hypothesis tested, the values
of the test statistic and the obtained pvalues.}
\item{levels}{Original levels of the variable \code{x}.}
\item{cluster}{A vector of integers (from 1:k) indicating the cluster to
which each curve is allocated.}
\item{centers}{An object containing the centroids
(mean of the curves pertaining to the same group).}
\item{curves}{An object containing the fitted curves for each population.}
}
\description{
Function for grouping cumulative incidence curves (CIF) based on
the k-means or k-medians algorithm. It returns the optimal number of
clusters and the assignment to the cluster which they belong to.
}
\details{
The adjustment methods include the Bonferroni correction ("bonferroni")
in which the p-values are multiplied by the number of comparisons.
Less conservative corrections are also included by Holm (1979) ('holm'),
Hochberg (1988) ('hochberg'), Hommel (1988) ('hommel'),
Benjamini & Hochberg (1995) ('BH' or its alias 'fdr'), and
Benjamini & Yekutieli (2001) ('BY'), respectively.
A pass-through option ('none') is also included.
}
\examples{
library(clustcif)
library(mstate)
data("ebmt2")

# Testing the number of groups
outtest <- clustcif(time = ebmt2$time, status = ebmt2$status, x = NULL,
                        kvector = NULL,  kbin = 20, nboot = 1,
                        algorithm = "kmeans", cluster = FALSE)
summary(outtest)

}
\author{
Nora M. Villanueva and Marta Sestelo.
}
